+++ front/lib/swr/workspaces.ts
@@ -4,7 +4,6 @@ import type { Fetcher } from "swr";
 import { emptyArray, fetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
 import type { GetWorkspaceResponseBody } from "@app/pages/api/w/[wId]";
 import type { GetWorkspaceFeatureFlagsResponseType } from "@app/pages/api/w/[wId]/feature-flags";
-import type { GetProvisioningStatusResponseBody } from "@app/pages/api/w/[wId]/provisioning-status";
 import type { GetSubscriptionsResponseBody } from "@app/pages/api/w/[wId]/subscriptions";
 import type { GetWorkspaceAnalyticsResponse } from "@app/pages/api/w/[wId]/workspace-analytics";
 import type { LightWorkspaceType, WhitelistableFeature } from "@app/types";
@@ -145,43 +144,5 @@ export function useFeatureFlags({
     isFeatureFlagsLoading: !error && !data,
     isFeatureFlagsError: error,
     hasFeature,
-  };
-}
-
-export function useProvisioningStatus({
-  workspaceId,
-  disabled,
-}: {
-  workspaceId: string;
-  disabled?: boolean;
-}) {
-  const provisioningStatusFetcher: Fetcher<GetProvisioningStatusResponseBody> =
-    fetcher;
-
-  const { data, error } = useSWRWithDefaults(
-    `/api/w/${workspaceId}/provisioning-status`,
-    provisioningStatusFetcher,
-    {
-      disabled,
-    }
-  );
-
-  const roleProvisioningStatus = useMemo(() => {
-    if (!data) {
-      return {
-        hasAdminGroup: false,
-        hasBuilderGroup: false,
-      };
-    }
-    return {
-      hasAdminGroup: data.hasAdminGroup,
-      hasBuilderGroup: data.hasBuilderGroup,
-    };
-  }, [data]);
-
-  return {
-    roleProvisioningStatus,
-    isProvisioningStatusLoading: !error && !data && !disabled,
-    isProvisioningStatusError: error,
   };
 }
